Networking Magic: some work done on client but then entered a frenzy of cognitive dissonance and am now in too shit a mood to code

+++ Projects/Prototypes/MultiClientServerTypesHelloer/src/MultiServerHelloer/multiserverhelloer.h
@@ -71,7 +71,7 @@ private:
     static void streamOutMagic(QDataStream *ds);
 signals:
     void d(const QString &);
-    void connectionHasBeenHelloedAndIsReadyForAction(QIODevice*, quint16);
+    void connectionHasBeenHelloedAndIsReadyForAction(QIODevice*, quint32);
 private slots:
     void handleNewClientConnected(QIODevice *newClient);
     void handleNewClientSentData();
